mcp-server/requirements.txt
mcp[http]>=1.8.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
uvicorn>=0.32.0
